+++ docs/CHANGELOG.md
@@ -46,6 +46,7 @@
     * `Hide Block Outline When GUI Is Hidden` (Enabled By Default)
     * `Fix Crash When Generating Certain Seeds` (Enabled By Default)
     * `Click Buttons On Mouse Down` (Enabled By Default)
+    * `3D Dropped Items` (Enabled By Default)
   * Existing Functionality (All Enabled By Default)
     * `Fix Screen Rendering When Hiding HUD`
     * `Sanitize Usernames`